DMA: Refactor
