+++ ProjectPSX/Devices/DMA/Channel.cs
@@ -1,0 +1,5 @@
+﻿namespace ProjectPSX.Devices; 
+public abstract class Channel {
+    public abstract void write(uint register, uint value);
+    public abstract uint load(uint regiter);
+}
